requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
psycopg2-binary==2.9.9
dj-database-url==2.1.0
python-decouple==3.8
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.0
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
python-telegram-bot==20.7
openai==1.3.5
pgvector==0.2.4
python-docx==1.1.0
requests==2.31.0
Pillow==10.1.0 